Add low level security on vacants and emails works

+++ src/main/java/mx/com/pandadevs/pibeapi/models/vacants/service/UserVacantService.java
@@ -1,5 +1,6 @@
 package mx.com.pandadevs.pibeapi.models.vacants.service;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import mx.com.pandadevs.pibeapi.models.auth.EmailService;
 import mx.com.pandadevs.pibeapi.models.processes.Process;
 import mx.com.pandadevs.pibeapi.models.processes.ProcessRepository;
@@ -13,6 +14,7 @@ import mx.com.pandadevs.pibeapi.models.vacants.entities.Vacant;
 import mx.com.pandadevs.pibeapi.models.vacants.mapper.UserVacantMapper;
 import mx.com.pandadevs.pibeapi.models.vacants.repository.UserVacantRepository;
 import mx.com.pandadevs.pibeapi.models.vacants.repository.VacantRepository;
+import mx.com.pandadevs.pibeapi.security.LogJwtService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,6 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -30,7 +33,6 @@ public class UserVacantService {
 
     private final UserVacantMapper mapper;
 
-
     public UserVacantService(UserVacantMapper mapper) {
         this.mapper = mapper;
     }
@@ -39,65 +41,78 @@ public class UserVacantService {
     private EmailService emailService;
     @Autowired
     private UserVacantRepository userVacantRepository;
-
     @Autowired
     private UserRepository userRepository;
     @Autowired
     private VacantRepository vacantRepository;
     @Autowired
     private ProcessRepository processRepository;
+    @Autowired
+    private LogJwtService logJwtService;
 
     private Optional<UserVacant> findByVacantAndUser(Long idUser, Integer idVacant) {
         return userVacantRepository.findByUser_IdAndVacant_Id(idUser, idVacant);
     }
 
     @Transactional(readOnly = true)
-    public List<UserVacantDto> getVacantsByUser(String username) {
-        try {
+    public List<UserVacantDto> getVacantsByUser(String username, String bearerToken) throws JsonProcessingException {
+        Map<String, String> auth = logJwtService.getUsernameAndRole(bearerToken);
+        if (auth.get("role").equals("ROLE_CANDIDATE") && auth.get("username").equals(username)) {
             return mapper.toUserVacantsDto(userVacantRepository.findAllByUser_Username(username));
-        } catch (Exception e) {
-            logger.error(e.getMessage());
         }
         return new ArrayList<>();
     }
 
     @Transactional(readOnly = true)
-    public List<VacantProcessDto> getUsersByVacant(Integer id) {
-        return mapper.toVacantsProcessDto(userVacantRepository.findAllByVacant_Id(id));
+    public List<VacantProcessDto> getUsersByVacant(Integer id, String bearerToken) throws JsonProcessingException {
+        Map<String, String> auth = logJwtService.getUsernameAndRole(bearerToken);
+        if (auth.get("role").equals("ROLE_RECRUITER")) {
+            return mapper.toVacantsProcessDto(userVacantRepository.findAllByVacant_Id(id));
+        }
+        return new ArrayList<>();
     }
 
     @Transactional
-    public Boolean applyToVacant(Integer id, String username) {
-        User user = userRepository.findByUsername(username);
-        Optional<Vacant> vacant = vacantRepository.findByIdAndActiveIsTrue(id);
-        Optional<Process> process = processRepository.findByIdAndActiveIsTrue(1);
-        Optional<UserVacant> userVacant = findByVacantAndUser(user.getId(), id);
-        if (user != null && vacant.isPresent() && process.isPresent() && !userVacant.isPresent()) {
-            emailService.sendEmailNewVacant(user, vacant.get());
-            userVacantRepository.save(new UserVacant(user, vacant.get(), process.get()));
-            return true;
+    public Boolean applyToVacant(Integer id, String username, String bearerToken) throws JsonProcessingException {
+        Map<String, String> auth = logJwtService.getUsernameAndRole(bearerToken);
+        if (auth.get("role").equals("ROLE_CANDIDATE") && auth.get("username").equals(username)) {
+            User user = userRepository.findByUsername(username);
+            Optional<Vacant> vacant = vacantRepository.findByIdAndActiveIsTrue(id);
+            Optional<Process> process = processRepository.findByIdAndActiveIsTrue(1);
+            Optional<UserVacant> userVacant = findByVacantAndUser(user.getId(), id);
+            if (vacant.isPresent() && process.isPresent() && userVacant.isEmpty()) {
+                emailService.sendEmailNewVacant(user, vacant.get());
+                userVacantRepository.save(new UserVacant(user, vacant.get(), process.get()));
+                return true;
+            }
         }
         return false;
     }
 
     @Transactional
-    public Boolean processToVacant(Integer id, ProcessDto processDto) {
-        Optional<UserVacant> userVacant = userVacantRepository.findById(id);
-        if (userVacant.isPresent()) {
-            userVacant.get().setProcess(processRepository.findByIdAndActiveIsTrue(processDto.getId()).get());
-            emailService.sendEmailCurrentlyProccess(userVacant.get(), userVacant.get().getProcess().getName().equals("Finalizado"));
-            userVacantRepository.save(userVacant.get());
-            return true;
+    public Boolean processToVacant(Integer id, ProcessDto processDto, String bearerToken) throws JsonProcessingException {
+        Map<String, String> auth = logJwtService.getUsernameAndRole(bearerToken);
+        if (auth.get("role").equals("ROLE_RECRUITER")) {
+            Optional<UserVacant> userVacant = userVacantRepository.findById(id);
+            if (userVacant.isPresent()) {
+                userVacant.get().setProcess(processRepository.findByIdAndActiveIsTrue(processDto.getId()).get());
+                emailService.sendEmailCurrentlyProccess(userVacant.get(), userVacant.get().getProcess().getName().equals("Finalizado"));
+                userVacantRepository.save(userVacant.get());
+                return true;
+            }
         }
         return false;
     }
 
     @Transactional
-    public Boolean deleteUserToVacant(Integer id) {
-        Optional<UserVacant> userVacant = userVacantRepository.findById(id);
-        if (userVacant.isPresent()) {
-            userVacantRepository.deleteById(id);
-            return true;
+    public Boolean deleteUserToVacant(Integer id, String bearerToken) throws JsonProcessingException {
+        Map<String, String> auth = logJwtService.getUsernameAndRole(bearerToken);
+        if (auth.get("role").equals("ROLE_RECRUITER")) {
+            Optional<UserVacant> userVacant = userVacantRepository.findById(id);
+            if (userVacant.isPresent()) {
+                userVacantRepository.deleteById(id);
+                return true;
+            }
         }
         return false;
     }
